Fix scrolling to a specific point in Safari
Fixes #4020
