+++ src/app/collections/PresentationNode.tsx
@@ -10,6 +10,7 @@ import classNames from 'classnames';
 import { expandIcon, collapseIcon, AppIcon } from '../shell/icons';
 import { deepEqual } from 'fast-equals';
 import { percent } from '../shell/filters';
+import { scrollToPosition } from 'app/dim-ui/scroll';
 
 interface Props {
   presentationNodeHash: number;
@@ -41,17 +42,11 @@ export default class PresentationNode extends React.Component<Props> {
       !deepEqual(this.lastPath, this.props.path)
     ) {
       const clientRect = this.headerRef.current.getBoundingClientRect();
-      const options: ScrollToOptions = {
+      scrollToPosition({
         top: window.scrollY + clientRect.top - 50,
         left: 0,
         behavior: 'smooth'
-      };
-      const isSmoothScrollSupported = 'scrollBehavior' in document.documentElement.style;
-      if (isSmoothScrollSupported) {
-        window.scroll(options);
-      } else {
-        window.scroll(options.top!, options.left!);
-      }
+      });
     }
     this.lastPath = this.props.path;
   }
